+++ web-gui/src/main.js
@@ -4,6 +4,8 @@ import router from "./router"
 
 // Plugins
 import Toast from "vue-toastification"
+import dayjs from "dayjs"
+import relativeTime from "dayjs/plugin/relativeTime"
 
 // Local styles
 import "../../web-gui-theme/dist/css/styles.css"
@@ -11,6 +13,16 @@ import "vue-toastification/dist/index.css"
 
 const app = createApp(App)
 app.use(router)
+
+// Global filters
+app.config.globalProperties.$filters = {
+  timeAgo(date) {
+    dayjs.extend(relativeTime)
+    return dayjs(date).fromNow()
+  },
+}
+
+app.config.globalProperties.$dayjs = dayjs
 
 // Use plugins
 app.use(Toast, {