+++ web-gui/src/views/SourceIPs.vue
@@ -6,7 +6,7 @@
         <div class="card-body">
           <nav class="nav nav-pills flex-column">
             <a class="nav-link active" aria-current="page" href="#">1 Hour</a>
-            <a class="nav-link" href="#">24 Hours</a>
+            <a class="nav-link" @click="ipAddressFilter(24)">24 Hours</a>
             <a class="nav-link" href="#">7 Days</a>
             <a class="nav-link" href="#">30 Days</a>
           </nav>
@@ -24,13 +24,17 @@
             <thead>
               <tr>
                 <th scope="col">IP Address</th>
+                <th class="text-end" scope="col">Last seen</th>
                 <th class="text-end" scope="col">Count</th>
               </tr>
             </thead>
             <tbody>
-              <tr v-for="item in ip_addresses" :key="item">
+              <tr v-for="item in ipAddressFilter" :key="item">
                 <td>{{ item.ip_address }}</td>
-                <td class="text-end">1576481</td>
+                <td class="text-end">
+                  {{ $filters.timeAgo(item.seen_last) }}
+                </td>
+                <td class="text-end">{{ item.seen_count }}</td>
               </tr>
             </tbody>
           </table>
@@ -47,7 +51,7 @@ export default {
   data() {
     return {
       name: "SourceIPs",
-      ip_addresses: null,
+      ip_addresses: [],
       ip_address_total: null,
     }
   },
@@ -64,7 +68,11 @@ export default {
             return Promise.reject(error)
           }
 
-          this.ip_addresses = data.ip_addresses
+          this.ip_addresses = data.ip_addresses.filter((ip_address) =>
+            this.$dayjs(ip_address.seen_last).isAfter(
+              this.$dayjs().subtract(1, "hours")
+            )
+          )
           this.ip_addresses_total = data.ip_addresses_total
         })
         .catch((error) => {
@@ -72,9 +80,54 @@ export default {
           toast.error("SentryPeer API error: " + error.message)
         })
     },
+    oneHour(ip_address) {
+      return this.$dayjs(ip_address.seen_last).isAfter(
+        this.$dayjs().subtract(1, "hours")
+      )
+    },
+    twentyFourHours(ip_address) {
+      return this.$dayjs(ip_address.seen_last).isAfter(
+        this.$dayjs().subtract(24, "hours")
+      )
+    },
+    sevenDays(ip_address) {
+      return this.$dayjs(ip_address.seen_last).isAfter(
+        this.$dayjs().subtract(7, "days")
+      )
+    },
+    thirtyDays(ip_address) {
+      return this.$dayjs(ip_address.seen_last).isAfter(
+        this.$dayjs().subtract(30, "days")
+      )
+    },
   },
   created() {
     this.getIPAddresses()
+  },
+  computed: {
+    ipAddressFilter(period) {
+      if (period === 1) {
+        return this.ip_addresses.filter((ip_address) =>
+          this.oneHour(ip_address)
+        )
+      }
+      if (period === 24) {
+        return this.ip_addresses.filter((ip_address) =>
+          this.twentyFourHours(ip_address)
+        )
+      }
+      if (period === 7) {
+        return this.ip_addresses.filter((ip_address) =>
+          this.sevenDays(ip_address)
+        )
+      }
+      if (period === 30) {
+        return this.ip_addresses.filter((ip_address) =>
+          this.thirtyDays(ip_address)
+        )
+      }
+      return this.ip_addresses
+    },
   },
 }
 </script>